# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job Board API",
        "description": "API for job listings with advanced filtering capabilities",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/jobs": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get filtered job listings",
                "description": "Returns a paginated list of job listings based on filter criteria",
                "operationId": "getJobs",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter expression (e.g. job_type = full-time AND languages HAS_ANY (PHP, JavaScript))",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of results per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Job"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "from": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "to": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid filter syntax",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Filter Error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid condition format"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 400
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/jobs/{id}": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get job details",
                "description": "Returns the details of a specific job listing",
                "operationId": "getJobById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Job"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Record not found."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Not Found"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Job": {
                "title": "Job Listing",
                "description": "Job listing model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Senior PHP Developer"
                    },
                    "description": {
                        "type": "string",
                        "example": "We are looking for an experienced PHP developer..."
                    },
                    "company_name": {
                        "type": "string",
                        "example": "Acme Inc"
                    },
                    "salary_min": {
                        "type": "number",
                        "format": "float",
                        "example": 60000
                    },
                    "salary_max": {
                        "type": "number",
                        "format": "float",
                        "example": 100000
                    },
                    "is_remote": {
                        "type": "boolean",
                        "example": true
                    },
                    "job_type": {
                        "type": "string",
                        "enum": [
                            "full-time",
                            "part-time",
                            "contract",
                            "freelance"
                        ],
                        "example": "full-time"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "draft",
                            "published",
                            "archived"
                        ],
                        "example": "published"
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-05-15T12:00:00Z"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-05-10T09:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-05-10T09:00:00Z"
                    },
                    "languages": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "name": {
                                    "type": "string",
                                    "example": "PHP"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "locations": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "city": {
                                    "type": "string",
                                    "example": "New York"
                                },
                                "state": {
                                    "type": "string",
                                    "example": "NY"
                                },
                                "country": {
                                    "type": "string",
                                    "example": "USA"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "name": {
                                    "type": "string",
                                    "example": "Web Development"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "attributes": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "name": {
                                    "type": "string",
                                    "example": "Years of Experience"
                                },
                                "type": {
                                    "type": "string",
                                    "enum": [
                                        "text",
                                        "number",
                                        "boolean",
                                        "date",
                                        "select"
                                    ],
                                    "example": "number"
                                },
                                "value": {
                                    "type": "mixed",
                                    "example": 5
                                },
                                "raw_value": {
                                    "type": "string",
                                    "example": "5"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Jobs",
            "description": "API endpoints for job listings"
        }
    ]
}